requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.1
django-allauth==0.38.0
django-crispy-forms==1.7.2
django-el-pagination==3.0.1
django-filters==0.2.1
django-import-export==1.1.0
django-markdown-deux==1.0.5
django-pagedown==1.0.6
django-rest-framework==0.1.0
django-simple-blog==0.3.0
django-simple-math-captcha==1.0.7
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
et-xmlfile==1.0.1
gunicorn==19.9.0
idna==2.7
jdcal==1.4
markdown2==2.3.6
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.8
Pillow==5.3.0
psycopg2==2.7.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
whitenoise==4.1.2
xlrd==1.1.0
xlwt==1.3.0
